Begin implementation of CV
Started implementing the CV procedure. Right now in the experiment loop,
what I am doing is running the CV procedure and storing all data. Then,
we can run the experiment for multiple seeds and hypers, and actually do
the tuning part of the CV offline. This way, we will have all the data
from the CV procedure offline and can explore that data more closely